requirements.txt
alembic==1.10.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
Babel @ file:///home/conda/feedstock_root/build_artifacts/babel_1677767029043/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1664823096650/work
brotlipy @ file:///D:/bld/brotlipy_1648854347500/work
certifi==2022.12.7
cffi @ file:///D:/bld/cffi_1656782940775/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
click @ file:///D:/bld/click_1666798499870/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
cryptography @ file:///D:/bld/cryptography_1637687432672/work
debugpy @ file:///C:/ci_310/debugpy_1642079916595/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
dnspython==2.3.0
dominate==2.7.0
email-validator==1.3.1
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1676592993069/work
flask-babel @ file:///home/conda/feedstock_root/build_artifacts/flask-babel_1675052342351/work
Flask-Bootstrap==3.3.7.1
Flask-Login @ file:///home/conda/feedstock_root/build_artifacts/flask-login_1658793187139/work
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-OpenID @ file:///home/conda/feedstock_root/build_artifacts/flask-openid_1636042720751/work
Flask-SQLAlchemy @ file:///home/conda/feedstock_root/build_artifacts/flask-sqlalchemy_1675194219027/work
Flask-WhooshAlchemy==0.56
Flask-WTF==1.1.1
flup==1.0.3
greenlet @ file:///C:/b/abs_47lk_w2ajq/croot/greenlet_1670513248400/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1672612343532/work
ipykernel @ file:///D:/bld/ipykernel_1655369313836/work
ipython @ file:///D:/bld/ipython_1677617262387/work
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1648147185463/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1654302431367/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1676579893731/work
jupyter_core @ file:///D:/bld/jupyter_core_1675110116927/work
langdetect @ file:///home/conda/feedstock_root/build_artifacts/langdetect_1622407545956/work
Mako==1.2.4
MarkupSafe @ file:///C:/ci/markupsafe_1654508036328/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1673482170163/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pbr @ file:///home/conda/feedstock_root/build_artifacts/pbr_1673485161019/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1677883256757/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1672682006896/work
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1666240235902/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
PySocks @ file:///D:/bld/pysocks_1661604991356/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/python-dotenv_1677227146928/work
python3-openid @ file:///home/conda/feedstock_root/build_artifacts/python3-openid_1612988842750/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1673864280276/work
pywin32==305.1
pyzmq @ file:///C:/ci/pyzmq_1657616000714/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1673863902341/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
SQLAlchemy @ file:///C:/Windows/Temp/abs_f8661157-660b-49bb-a790-69ab9f3b8f7c8a8s2psb/croots/recipe/sqlalchemy_1657867864564/work
sqlalchemy-migrate==0.13.0
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1663990453245/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
Tempita==0.5.3.dev0
tornado @ file:///D:/bld/tornado_1656937966227/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1665144421445/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1678635778344/work
visitor==0.1.3
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1676411946679/work
Whoosh==2.7.4
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
wincertstore==0.2
WTForms @ file:///home/conda/feedstock_root/build_artifacts/wtforms_1640306863281/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
